Move tests for PathWalker to a separate test file
